Move CanIsolate CompletableFuture instance to 2.12+ only
